is.SharedData {crosstalk}	R Documentation

Check if an object is SharedData

Description

Check if an object is an instance of SharedData or not.

Usage

is.SharedData(x)

Arguments

x
The object that may or may not be an instance of SharedData

Value

logical

[Package crosstalk version 1.2.0 Index]