filter_select {crosstalk}	R Documentation

Categorical filter controls

Description

Creates a select box or list of checkboxes, for filtering a SharedData object based on categorical data.

Usage

filter_select(id, label, sharedData, group, allLevels = FALSE, multiple = TRUE)

filter_checkbox(
  id,
  label,
  sharedData,
  group,
  allLevels = FALSE,
  inline = FALSE,
  columns = 1
)

Arguments

id
An HTML element ID; must be unique within the web page

label
A human-readable label

sharedData
SharedData object with the data to filter

group
A one-sided formula whose values will populate this select box. Generally this should be a character or factor column; if not, it will be coerced to character.

allLevels
If the vector described by group is factor-based, should all the levels be displayed as options, or only ones that are present in the data?

multiple
Can multiple values be selected?

inline
If TRUE, render checkbox options horizontally instead of vertically.

columns
Number of columns the options should be arranged into.

Examples

## Only run examples in interactive R sessions
if (interactive()) {

sd <- SharedData$new(chickwts)
filter_select("feedtype", "Feed type", sd, "feed")

}

[Package crosstalk version 1.2.1 Index]