filter_slider {crosstalk}	R Documentation

Range filter control

Description

Creates a slider widget that lets users filter observations based on a range of values.

Usage

filter_slider(
  id,
  label,
  sharedData,
  column,
  step = NULL,
  round = FALSE,
  ticks = TRUE,
  animate = FALSE,
  width = NULL,
  sep = ",",
  pre = NULL,
  post = NULL,
  timeFormat = NULL,
  timezone = NULL,
  dragRange = TRUE,
  min = NULL,
  max = NULL
)

animation_options(
  interval = 1000,
  loop = FALSE,
  playButton = NULL,
  pauseButton = NULL
)

Arguments

id
An HTML element ID; must be unique within the web page

label
A human-readable label

sharedData
SharedData object with the data to filter

column
A one-sided formula whose values will be used for this slider. The column must be of type Date, POSIXt, or numeric.

step
Specifies the interval between each selectable value on the slider (if NULL, a heuristic is used to determine the step size). If the values are dates, step is in days; if the values are times (POSIXt), step is in seconds.

round
TRUE to round all values to the nearest integer; FALSE if no rounding is desired; or an integer to round to that number of decimal places (for example, 1 will round to the nearest 0.1, and -2 will round to the nearest 100). Any rounding will be applied after snapping to the nearest step.

ticks
FALSE to hide tick marks, TRUE to show them according to some simple heuristics.

animate
TRUE to show simple animation controls with default settings; FALSE not to; or a custom settings list, such as those created using animationOptions.

width
The width of the slider control (see validateCssUnit for valid formats)

sep
Separator between thousands places in numbers.

pre
A prefix string to put in front of the value.

post
A suffix string to put after the value.

timeFormat
Only used if the values are Date or POSIXt objects. A time format string, to be passed to the Javascript strftime library. See https://github.com/samsonjs/strftime for more details. The allowed format specifications are very similar, but not identical, to those for R's strftime function. For Dates, the default is "%F" (like "2015-07-01"), and for POSIXt, the default is "%F %T" (like "2015-07-01 15:32:10").

timezone
Only used if the values are POSIXt objects. A string specifying the time zone offset for the displayed times, in the format "+HHMM" or "-HHMM". If NULL (the default), times will be displayed in the browser's time zone. The value "+0000" will result in UTC time.

dragRange
This option is used only if it is a range slider (with two values). If TRUE (the default), the range can be dragged. In other words, the min and max can be dragged together. If FALSE, the range cannot be dragged.

min
The leftmost value of the slider. By default, set to the minimal number in input data.

max
The rightmost value of the slider. By default, set to the maximal number in input data.

interval
The interval, in milliseconds, between each animation step.

loop
TRUE to automatically restart the animation when it reaches the end.

playButton
Specifies the appearance of the play button. Valid values are a one-element character vector (for a simple text label), an HTML tag or list of tags (using tag and friends), or raw HTML (using HTML).

pauseButton
Similar to playButton, but for the pause button.

Examples

## Only run examples in interactive R sessions
if (interactive()) {

sd <- SharedData$new(mtcars)
filter_slider("mpg", "Miles per gallon", sd, "mpg")

}

[Package crosstalk version 1.2.0 Index]